+++ src/dataset.rs
@@ -4,8 +4,9 @@ pub mod iris;
 /// This module provides access to the diabetes dataset
 pub mod diabetes;
 
-/// This module provides access to the wine quality dataset (red wine and white wine)
-pub mod wine_quality;
-
 /// This module provides access to the raw data
 pub mod raw_data;
+/// This module provides access to the titanic dataset
+pub mod titanic;
+/// This module provides access to the wine quality dataset (red wine and white wine)
+pub mod wine_quality;