+++ src/dataset/raw_data/diabetes_raw.rs
@@ -1,9 +1,10 @@
 /// Loads raw diabetes data and its corresponding headers as static strings
 ///
 /// # Returns
-/// A tuple containing:
-/// - `&'static str`: The column headers
-/// - `&'static str`: The raw dataset rows
+///
+/// * A tuple containing:
+///     - `&'static str`: The column headers
+///     - `&'static str`: The raw dataset rows
 pub fn load_diabetes_raw_data() -> (&'static str, &'static str) {
     (
         // diabetes_data_headers_raw