+++ changelog.md
@@ -11,6 +11,7 @@ Please view [SomeB1oody/RustyML](https://github.com/SomeB1oody/RustyML) for more
 
 ### Added
 - Add `load_titanic_raw_data` function to provide static raw Titanic dataset and headers
+- Add cost calculation and reporting for ML models
 
 ## [v0.7.0] - 2025-9-21 (UTC-7)
 ### Added